Fix incorrect value in bad character table for first character in needle
Previously, if the first character in the needle did not occur anywhere
else in the needle, the value of the bad character table for that
character was was the length of the needle when it should have been one
less than the length of the needle.

Thanks to @RalphPastel972 for noticing the issue!